add post check logic, some refactoring

+++ controllers/PostController.js
@@ -1,4 +1,5 @@
 import PostModel from "../models/Post.js";
+import PostCheckModel from "../models/PostCheck.js";
 import UserModel from "../models/User.js";
 
 import { fileURLToPath } from "url";
@@ -12,7 +13,7 @@ const __dirname = path.dirname(__filename);
 
 export const create = async (req, res) => {
   try {
-    const user = req.userId;
+    const user = req.body.userId;
     const doc = new PostModel({
       title: req.body.title,
       text: req.body.text,
@@ -52,6 +53,36 @@ export const create = async (req, res) => {
   }
 };
 
+export const createCheck = async (req, res) => {
+  try {
+    const user = req.userId;
+
+    const postVerify = PostModel.find({ text: req.body.text });
+
+    if (postVerify.length > 0) {
+      return res.json({
+        message: "This post is not unique",
+      });
+    }
+
+    const doc = new PostCheckModel({
+      title: req.body.title,
+      text: req.body.text,
+      tags: req.body.tags,
+      imageUrl: req.body.imageUrl,
+      user,
+    });
+
+    const postCheck = await doc.save();
+    res.json(postCheck);
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: "Something wrong when create post",
+    });
+  }
+};
+
 ///////////////////////// - read
 
 export const getAll = async (req, res) => {
@@ -62,6 +93,34 @@ export const getAll = async (req, res) => {
     console.log(err);
     res.status(500).json({
       message: "Failed to get posts",
+    });
+  }
+};
+
+export const getAllCheck = async (req, res) => {
+  try {
+    const posts = await PostCheckModel.find().populate("user").exec();
+
+    res.json(posts);
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: "Failed to get posts to check",
+    });
+  }
+};
+
+export const getOneCheck = async (req, res) => {
+  try {
+    const postId = req.params.id;
+
+    const post = await PostCheckModel.findById(postId).populate("user");
+
+    res.json(post);
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: "Failed to get post",
     });
   }
 };
@@ -199,6 +258,23 @@ export const getPostsByUserId = async (req, res) => {
   }
 };
 
+export const getCheckPostsByUserId = async (req, res) => {
+  try {
+    const userId = req.params.id;
+
+    const posts = await PostCheckModel.find({
+      user: userId,
+    }).exec();
+
+    res.json(posts);
+  } catch (err) {
+    console.warn(err);
+    res.status(500).json({
+      message: "Failed to get user posts",
+    });
+  }
+};
+
 ///////////////////////// - update
 
 export const update = async (req, res) => {
@@ -207,7 +283,7 @@ export const update = async (req, res) => {
 
     const { imageUrl } = await PostModel.findById(postId);
 
-    if (imageUrl) {
+    if (imageUrl && imageUrl !== req.body.imageUrl) {
       const filePath = path.join(__dirname, "..", imageUrl);
       if (fs.existsSync(filePath)) {
         fs.unlinkSync(filePath);
@@ -240,6 +316,45 @@ export const update = async (req, res) => {
   }
 };
 
+export const updateCheck = async (req, res) => {
+  try {
+    const postId = req.params.id;
+
+    const { imageUrl } = await PostCheckModel.findById(postId);
+
+    if (imageUrl && imageUrl !== req.body.imageUrl) {
+      const filePath = path.join(__dirname, "..", imageUrl);
+      if (fs.existsSync(filePath)) {
+        fs.unlinkSync(filePath);
+      } else {
+        console.log("File delete (update post)");
+      }
+    }
+
+    await PostCheckModel.updateOne(
+      {
+        _id: postId,
+      },
+      {
+        title: req.body.title,
+        text: req.body.title,
+        tags: req.body.tags,
+        imageUrl: req.body.imageUrl,
+        isVerifyEdit: req.body.isVerifyEdit,
+        comment: req.body.comment,
+      }
+    );
+
+    res.json({
+      success: true,
+    });
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: "Failed to update post",
+    });
+  }
+};
 ///////////////////////// - delete
 
 export const remove = async (req, res) => {
@@ -248,9 +363,9 @@ export const remove = async (req, res) => {
     const userId = req.params.userId;
 
     const { imageUrl } = await PostModel.findById(postId);
-    const filename = imageUrl.slice(14);
 
     if (imageUrl) {
+      const filename = imageUrl.slice(14);
       const filePath = path.join(__dirname, "../uploads/post", filename);
       if (fs.existsSync(filePath)) {
         fs.unlinkSync(filePath, (err) => {
@@ -289,6 +404,41 @@ export const remove = async (req, res) => {
           console.log(err);
           return res.status(500).json({
             message: "Failed to delete post",
+          });
+        }
+
+        if (!doc) {
+          return res.status(404).json({
+            message: "Post has not found",
+          });
+        }
+
+        res.json({
+          success: true,
+        });
+      }
+    );
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({
+      message: "Failed to delete post",
+    });
+  }
+};
+
+export const removeCheck = async (req, res) => {
+  try {
+    const postId = req.params.id;
+
+    PostCheckModel.findOneAndDelete(
+      {
+        _id: postId,
+      },
+      (err, doc) => {
+        if (err) {
+          console.log(err);
+          return res.status(500).json({
+            message: "Problem to delete post",
           });
         }
